% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junctions_to_anchmm.R
\name{junctions_to_anchmm}
\alias{junctions_to_anchmm}
\title{converts output from simulations of the junctions package to input for
Ancestry_HMM}
\usage{
junctions_to_anchmm(sim_data)
}
\arguments{
\item{sim_data}{output from junctions::sim_phased_unphased}
}
\value{
list with posterior probabilities, age estimate, and likelihood
}
\description{
converts output from simulations of the junctions package to input for
Ancestry_HMM
}
