% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junctions_to_anchmm.R
\name{junctions_to_anchmm_imperfect}
\alias{junctions_to_anchmm_imperfect}
\title{converts output from simulations of the junctions package to input for
Ancestry_HMM}
\usage{
junctions_to_anchmm_imperfect(sim_data, d_value)
}
\arguments{
\item{sim_data}{output from junctions::sim_phased_unphased}

\item{d_value}{average discriminatory value of the markers used, where the
discriminatory value is given by: p_A / (p_A + p_a), e.g. the d value
indicates the probability of getting the ancestry right.}
}
\value{
list with posterior probabilities, age estimate, and likelihood
}
\description{
converts output from simulations of the junctions package to input for
Ancestry_HMM
}
