% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_to_anchmm_using_data.R
\name{junctions_to_anchmm_using_data}
\alias{junctions_to_anchmm_using_data}
\title{converts output from simulations of the junctions package to input for
Ancestry_HMM}
\usage{
junctions_to_anchmm_using_data(sim_data, emp_data)
}
\arguments{
\item{sim_data}{output from junctions::sim_phased_unphased}

\item{emp_data}{empirical data.}
}
\value{
list with posterior probabilities, age estimate, and likelihood
}
\description{
converts output from simulations of the junctions package to input for
Ancestry_HMM
}
