\name{rancestry-package}
\alias{rancestry-package}
\alias{rancestry}
\docType{package}
\title{\packageTitle{rancestry}}
\description{\packageDescription{rancestry}}
\section{Package Content}{\packageIndices{rancestry}}
\author{\packageAuthor{rancestry}}
\section{Maintainer}{\packageMaintainer{rancestry}}
\keyword{package}
