% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ancestry_hmm.R
\name{run_ancestry_hmm}
\alias{run_ancestry_hmm}
\title{run Ancestry_HMM}
\usage{
run_ancestry_hmm(
  sample_matrix,
  genetic_data,
  cmd_line_options = numeric(0),
  viterbi = FALSE,
  pulse1 = numeric(0),
  pulse2 = numeric(0),
  ancestry_proportions = c(2, 0.5, 0.5),
  use_genetic_data = TRUE
)
}
\arguments{
\item{sample_matrix}{matrix with sample names and ploidy}

\item{genetic_data}{matrix with the genetic data (see manual)}

\item{cmd_line_options}{cmd line entries in a vector, where each element
separated by a space is in a different position, e.g. c("-i", "input.txt")}

\item{viterbi}{should viterbi output be generated?}

\item{pulse1}{if no cmd_line_options are given, a pulse can be added manually
with 3 properties: c(ancestor, time, proportion)}

\item{pulse2}{if no cmd_line_options are given, a pulse can be added manually
with 3 properties: c(ancestor, time, proportion)}

\item{ancestry_proportions}{ancestry proportions}

\item{use_genetic_data}{is the genetic data in reads, or alleles?}
}
\value{
list with posterior probabilities, age estimate, and likelihood
}
\description{
run Ancestry_HMM
}
